challenge1/requirements.txt
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0